move object_store logic to metastore crate
